10_Pipeline_Orchestration/2_airflow/airflow_essential/requirements.txt
apache-airflow
apache-airflow-providers-apache-spark